Combined all shaders in one es6 module.
