--- app/modes/disco.js
@@ -1,9 +1,0 @@
-'use strict';
-
-var mode = angular.module('sos.modes');
-mode.factory('modeDisco', function(audioDisco, $log) {
-  return new ShaderMode({ id: 'modeDisco',
-                          title: 'Disco',
-                          audio: audioDisco,
-                          pixelShaderName: 'discoFrag' });
-});
