+++ app/modes/midi.js
@@ -2,15 +2,15 @@
 
 var mode = angular.module('sos.modes');
 mode.factory('modeMIDI', function($log) {
-  
+
   var mode = {};
-  
+
   mode.id = "modeMIDI";
   mode.title = "MIDI Keyboard Mode";
   mode.socket = {};
 
   mode.init = function($scope) {
-    
+
     MIDI.loadPlugin({
       instrument: "acoustic_grand_piano", // or the instrument code 1 (aka the default)
       onsuccess: function() { }
@@ -18,7 +18,6 @@ mode.factory('modeMIDI', function($log) {
 
     mode.socket = io.connect('http://localhost:1337');
     mode.socket.on('message', function(data){
-      console.log(data);
 
       var note = data.note;
 
@@ -41,15 +40,15 @@ mode.factory('modeMIDI', function($log) {
     circle.x = circle.y = 50;
     //Add Shape instance to stage display list.
     $scope.stage.addChild(circle);
-  }
+  };
 
   mode.update = function($scope) {
     // no-op
-  }
-  
+  };
+
   mode.deinit = function($scope) {
     mode.socket.close();
-  }
+  };
 
   return mode;
 });
