+++ app/modes/slowclap.js
@@ -1,39 +1,41 @@
 'use strict';
 
+let Mode = require('../mode');
+
 var mode = angular.module('sos.modes');
 mode.factory('modeSlowClap', function($log) {
-	
-	var count = 0;
-	var mode = new Mode("modeSlowClap", "Slow Clap");
-	var loader = PIXI.loader;
-	
-	mode.spritesheetJSON = null;
-	mode.movie = null;
-	mode.filter = null;
-	
-	mode.init = function(parentScope) {
-	
-		mode.setParentScope(parentScope);
-		mode.container = new PIXI.Container();
-    
+
+  var count = 0;
+  var mode = new Mode("modeSlowClap", "Slow Clap");
+  var loader = PIXI.loader;
+
+  mode.spritesheetJSON = null;
+  mode.movie = null;
+  mode.filter = null;
+
+  mode.init = function(parentScope) {
+
+    mode.setParentScope(parentScope);
+    mode.container = new PIXI.Container();
+
     if(!loader.resources.spritesheet) {
-      PIXI.loader.add('spritesheet', 'media/slow-clap.json').load(mode.initAnimation);  
+      PIXI.loader.add('spritesheet', 'media/slow-clap.json').load(mode.initAnimation);
     } else {
       mode.initAnimation();
     }
-  }
-	
-	mode.initAnimation = function() {
-		
-		// create an array of textures from an image path
-		var frames = [];
+  };
+
+  mode.initAnimation = function() {
+
+    // create an array of textures from an image path
+    var frames = [];
 
     for (var i = 0; i < 7; i++) {
-        var val = i < 10 ? '0' + i : i;
+      var val = i < 10 ? '0' + i : i;
 
-        // magically works since the spritesheet was loaded with the pixi loader
-        var spritesheet
-        frames.push(PIXI.Texture.fromFrame('citizen-kane-clapping_0' + val + '.png'));
+      // magically works since the spritesheet was loaded with the pixi loader
+      var spritesheet;
+      frames.push(PIXI.Texture.fromFrame('citizen-kane-clapping_0' + val + '.png'));
     }
 
 
@@ -46,23 +48,23 @@ mode.factory('modeSlowClap', function($log) {
      */
     mode.movie.position.set(0);
     mode.movie.anchor.set(0);
-    
+
     var newScale = new PIXI.Point(mode.parentScope.getWidthScaleFactor(mode.movie.width), mode.parentScope.getHeightScaleFactor(mode.movie.height));
 
     mode.movie.scale = newScale;
     mode.movie.animationSpeed = 0.5;
     mode.movie.play();
-		mode.container.addChild(mode.movie);
-		
-		// set up color filter
-		mode.filter = new PIXI.filters.ColorMatrixFilter();
-		mode.container.filters = [mode.filter];
-		
-		mode.renderID = requestAnimationFrame(mode.update);	
-	}
-	
-	mode.update = function() {
-		
+    mode.container.addChild(mode.movie);
+
+    // set up color filter
+    mode.filter = new PIXI.filters.ColorMatrixFilter();
+    mode.container.filters = [mode.filter];
+
+    mode.renderID = requestAnimationFrame(mode.update);
+  };
+
+  mode.update = function() {
+
     var matrix = mode.filter.matrix;
 
     count += 0.1;
@@ -72,15 +74,15 @@ mode.factory('modeSlowClap', function($log) {
     matrix[4] = Math.sin(count / 3) * 2;
     matrix[5] = Math.sin(count / 2);
     matrix[6] = Math.sin(count / 4);
-		
-		
-		mode.parentScope.pixijs.renderer.render(mode.container);
-		requestAnimationFrame(mode.update);			
-	}
 
-	mode.deinit = function() {
-		cancelAnimationFrame(self.renderID);
-	} 
-	
-	return mode;	
+
+    mode.parentScope.pixijs.renderer.render(mode.container);
+    requestAnimationFrame(mode.update);
+  };
+
+  mode.deinit = function() {
+    cancelAnimationFrame(this.renderID);
+  };
+
+  return mode;
 });
