--- app/modes/tunnel.js
@@ -1,8 +1,0 @@
-'use strict';
-
-var mode = angular.module('sos.modes');
-mode.factory('modeTunnel', function($log) {
-  return new ShaderMode({ id: 'modeTunnel',
-                          title: 'Tunnel',
-                          pixelShaderName: 'tunnelFrag' });
-});
