--- app/modes/worms.js
@@ -1,8 +1,0 @@
-'use strict';
-
-var mode = angular.module('sos.modes');
-mode.factory('modeWorms', function($log) {
-  return new ShaderMode({ id: 'modeWorms',
-                          title: 'Worms',
-                          pixelShaderName: 'wormsFrag' });
-});
