+++ bower.json
@@ -12,7 +12,6 @@
     "midi": "*",
     "socket.io-client": "~1.3.6",
     "tweenjs": "*",
-    "proton": "*",
     "pako": "*",
     "three.js": "~0.71.0",
     "keyboardjs": "~2.1.1",
