+++ package.json
@@ -12,6 +12,7 @@
     "gulp-sourcemaps": "latest",
     "gulp-babel": "latest",
     "gulp-concat": "latest",
+    "gulp-connect": "latest",
     "bower": "^1.3.1",
     "http-server": "^0.6.1",
     "socket.io-client": "^1.3.6"
